backend/requirements.txt
requests==2.26.0
Django==3.2.18
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
pytest==6.2.4
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.9.5
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
isort==5.10.1
Pillow==9.1.1
pyflakes==2.4.0
PyJWT==2.4.0
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python-dotenv==0.20.0
djoser==2.1.0
drf-extra-fields==3.4.0
django-colorfield==0.8.0